Store coefficient textures as Vec<f32>
This means we don't need to store the alpha values if the texture has no
alpha channel.

+++ src/scene.rs
@@ -88,7 +88,7 @@ impl Scene {
                             other => panic!("Texture format {:?} is not implemented", other),
                         };
 
-                        Texture::new(pixels, i.width, i.height, format)
+                        Texture::new(pixels, i.width as usize, i.height as usize, format)
                     })
                     .collect::<Vec<Texture>>();
